More separation and typing
